Refact: Move scopes into the spring.scope package
